requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
trafilatura>=1.12.2
beautifulsoup4>=4.12.3
readability-lxml>=0.8.1
lxml>=5.2.2
pydantic>=2.8.2
rich>=13.8.1
pydantic-settings>=2.5.2
openai>=1.46.0
duckduckgo-search>=6.3.7
serpapi>=0.1.5
click>=8.1.7
python-docx>=1.1.2
